zabbix_api/zapi/requirements.txt
djangorestframework
redis
pytz
requests
